refactor: move OpenItem to airflow::model
